Move attribute value type enum to separate file
