+++ src/products/entities/attribute-value-type.enum.ts
@@ -1,0 +1,6 @@
+export enum AttributeValueType {
+  String = 'string',
+  Number = 'number',
+  Boolean = 'boolean',
+  Color = 'color',
+}
